+++ example/lib/main.dart
@@ -1,73 +1,71 @@
-import 'package:ccppflutterplugin/ccppflutterplugin.dart';
 import 'package:flutter/material.dart';
 import 'dart:async';
 
-void main() async {
+import 'package:flutter/services.dart';
+import 'package:flutter_ccpp/flutter_ccpp.dart';
+import 'package:flutter_ccpp/pigeon.dart';
+
+void main() {
   runApp(MyApp());
 }
 
 class MyApp extends StatefulWidget {
   @override
-  _MyAppState createState() {
-    return _MyAppState();
-  }
+  _MyAppState createState() => _MyAppState();
 }
 
 class _MyAppState extends State<MyApp> {
+  String _platformVersion = 'Unknown';
+
   @override
   void initState() {
     super.initState();
-    initCcppPlugin();
+    initPlatformState();
   }
 
   // Platform messages are asynchronous, so we initialize in an async method.
-  Future<void> initCcppPlugin() async {
-    await CcppFlutterPlugin.initialize(
-      isSandbox: true,
-    );
+  Future<void> initPlatformState() async {
+    String platformVersion;
+    // Platform messages may fail, so we use a try/catch PlatformException.
+    // We also handle the message potentially returning null.
+    try {
+      FlutterCcpp.setup(CcppEnvironment.sandbox);
+      platformVersion = 'Initialized';
+      // Try make a new payment!
+      var payment = await FlutterCcpp.makePanCreditCardPayment(
+        "kSAops9Zwhos8hSTSeLTUa+y/Hc8FJ1w4/jAGUqhUcMU6mDXSirHZjqtn/wEgJi1w3nONLuU7kSTUYkknSfwRwNsxt7JB+kB+HYMCu5KbquzbbvRgzQEj455GM01dz8e",
+        "4111111111111111",
+        2,
+        2022,
+        "123",
+        true
+      );
+      print(payment.responseCode);
+      print(payment.redirectUrl);
+      print(payment.error);
+    } catch(e) {
+      platformVersion = e.toString();
+    }
+
+    // If the widget was removed from the tree while the asynchronous platform
+    // message was in flight, we want to discard the reply rather than calling
+    // setState to update our non-existent appearance.
+    if (!mounted) return;
+
+    setState(() {
+      _platformVersion = platformVersion;
+    });
   }
 
   @override
   Widget build(BuildContext context) {
-    var paymentToken = 'PAYMENT_TOKEN';
-    var securityCode = '123';
     return MaterialApp(
       home: Scaffold(
         appBar: AppBar(
           title: const Text('Plugin example app'),
         ),
         body: Center(
-          child: Column(
-            children: <Widget>[
-              RaisedButton(
-                child: Text('pay'),
-                onPressed: () async {
-                  var response = await CcppFlutterPlugin.paymentWithCreditCard(
-                    paymentToken: paymentToken,
-                    creditCardNumber: '4111111111111111',
-                    expiryMonth: 7,
-                    expiryYear: 2024,
-                    securityCode: securityCode,
-                    storeCard: true,
-                  );
-                  print('invoiceNo: ${response.invoiceNo}');
-                  print('error: ${response.errorMessage}');
-                },
-              ),
-              RaisedButton(
-                child: Text('pay with token'),
-                onPressed: () async {
-                  var response = await CcppFlutterPlugin.paymentWithToken(
-                    paymentToken: paymentToken,
-                    cardToken: 'CARD_TOKEN',
-                    securityCode: securityCode,
-                  );
-                  print('invoiceNo: ${response.invoiceNo}');
-                  print('error: ${response.errorMessage}');
-                },
-              ),
-            ],
-          ),
+          child: Text('Running on: $_platformVersion\n'),
         ),
       ),
     );