+++ ios/Classes/CcppFlutterPlugin.h
@@ -1,4 +1,0 @@
-#import <Flutter/Flutter.h>
-
-@interface CcppFlutterPlugin : NSObject<FlutterPlugin>
-@end