+++ CHANGELOG.md
@@ -1,3 +1,36 @@
-## 0.1.0
+## 0.0.7
 
-* TODO: Describe initial release.
+* Added `setAutoResetSSLSocket(bool)` method.
+* `setAutoResetSSLSocket(true)` will make sure Android reset the SSL connection
+every time the request has completed.
+
+## 0.0.6
+
+* Fixed QR Code payment for iOS
+
+## 0.0.5
+
+* Fixed QR Code payment for Android
+
+## 0.0.4
+
+* Added static utliity method to callable to map the response code upon request.
+
+## 0.0.3
+
+* Allow QR Payment to configure payment channel as per [DOC](https://developer.2c2p.com/docs/reference-payment-channels#section-qr-payment-qrc-csqr-thqr-sgqr)
+
+## 0.0.2
+
+* Added progard for Android, so that it can be published.
+* Added additional payment options
+  * [Payment with Installment via CardToken](https://developer.2c2p.com/v4.0.2/docs/sdk-installment-payment-plan)
+  * [Payment with Installment via PAN](https://developer.2c2p.com/v4.0.2/docs/sdk-installment-payment-plan)
+  * [Payment with CreditCard](https://developer.2c2p.com/v4.0.2/docs/sdk-method-qr-payment)
+
+## 0.0.1
+
+* Added 2 payment options
+  * [SimplePayment](https://developer.2c2p.com/v4.0.2/docs/sdk-payment-authorization)
+  * [Payment with CardToken](https://developer.2c2p.com/v4.0.2/docs/sdk-payment-with-card-token)
+